separate interceptor, model
