+++ src/sdk/admin/mod.rs
@@ -9,12 +9,13 @@ use crate::genpb::cerbos::{
         DisablePolicyRequest, EnablePolicyRequest, GetPolicyRequest, GetSchemaRequest,
         InspectPoliciesRequest, ListPoliciesRequest, ListSchemasRequest, ReloadStoreRequest,
     },
-    response::v1::{InspectPoliciesResponse, ListAuditLogEntriesResponse},
+    response::v1::InspectPoliciesResponse,
     schema::v1::Schema,
 };
-use anyhow::{Context, Result};
+use anyhow::{anyhow, Context, Result};
 use base64::prelude::{Engine as _, BASE64_STANDARD};
-use std::time::SystemTime;
+use model::{FilterOptions, PolicySet, SchemaSet};
+use std::time::Duration;
 use tonic::metadata::MetadataValue;
 use tonic::service::interceptor::InterceptedService;
 use tonic::service::Interceptor;
@@ -26,214 +27,13 @@ use super::CerbosClientOptions;
 const ADD_POLICY_BATCH_SIZE: usize = 10;
 const ADD_SCHEMA_BATCH_SIZE: usize = 10;
 
-/// Admin client for the Cerbos Admin API
-pub struct CerbosAdminClient {
-    client: CerbosAdminServiceClient<InterceptedService<Channel, BasicAuthInterceptor>>,
-}
-
-/// Policy set container for managing multiple policies
-#[derive(Debug, Clone, Default)]
-pub struct PolicySet {
-    policies: Vec<Policy>,
-}
-
-/// Schema set container for managing multiple schemas
-#[derive(Debug, Clone, Default)]
-pub struct SchemaSet {
-    schemas: Vec<Schema>,
-}
-
-/// Options for filtering policies and schemas
-#[derive(Debug, Clone, Default)]
-pub struct FilterOptions {
-    pub policy_ids: Vec<String>,
-    pub include_disabled: bool,
-    pub name_regexp: Option<String>,
-    pub scope_regexp: Option<String>,
-    pub version_regexp: Option<String>,
-}
-
-/// Audit log types
-#[derive(Debug, Clone, Copy)]
-pub enum AuditLogType {
-    Access,
-    Decision,
-}
-
-/// Options for audit log retrieval
-#[derive(Debug, Clone)]
-pub struct AuditLogOptions {
-    pub start_time: Option<SystemTime>,
-    pub end_time: Option<SystemTime>,
-    pub lookup: Option<String>,
-    pub tail: Option<u32>,
-    pub log_type: AuditLogType,
-}
-
-/// Audit log entry wrapper
-#[derive(Debug, Clone)]
-pub struct AuditLogEntry {
-    response: ListAuditLogEntriesResponse,
-    error: Option<String>,
-}
+pub mod model;
 
 /// Basic authentication credentials
 #[derive(Debug, Clone)]
 pub struct BasicAuth {
     username: String,
     password: String,
-}
-
-impl PolicySet {
-    /// Create a new empty policy set
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Add policies to the set
-    pub fn add_policies(&mut self, policies: Vec<Policy>) -> &mut Self {
-        self.policies.extend(policies);
-        self
-    }
-
-    /// Add a single policy to the set
-    pub fn add_policy(&mut self, policy: Policy) -> &mut Self {
-        self.policies.push(policy);
-        self
-    }
-
-    /// Get all policies in the set
-    pub fn get_policies(&self) -> &[Policy] {
-        &self.policies
-    }
-
-    /// Get the size of the policy set
-    pub fn size(&self) -> usize {
-        self.policies.len()
-    }
-
-    /// Validate the policy set
-    pub fn validate(&self) -> Result<()> {
-        if self.policies.is_empty() {
-            anyhow::bail!("empty policy set");
-        }
-        Ok(())
-    }
-}
-
-impl SchemaSet {
-    /// Create a new empty schema set
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Add schemas to the set
-    pub fn add_schemas(&mut self, schemas: Vec<Schema>) -> &mut Self {
-        self.schemas.extend(schemas);
-        self
-    }
-
-    /// Add a single schema to the set
-    pub fn add_schema(&mut self, schema: Schema) -> &mut Self {
-        self.schemas.push(schema);
-        self
-    }
-
-    /// Get all schemas in the set
-    pub fn get_schemas(&self) -> &[Schema] {
-        &self.schemas
-    }
-
-    /// Get the size of the schema set
-    pub fn size(&self) -> usize {
-        self.schemas.len()
-    }
-}
-
-impl FilterOptions {
-    /// Create new filter options
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Set policy IDs filter
-    pub fn with_policy_ids(mut self, ids: Vec<String>) -> Self {
-        self.policy_ids = ids;
-        self
-    }
-
-    /// Include disabled policies/schemas
-    pub fn with_include_disabled(mut self, include: bool) -> Self {
-        self.include_disabled = include;
-        self
-    }
-
-    /// Set name regexp filter
-    pub fn with_name_regexp(mut self, regexp: String) -> Self {
-        self.name_regexp = Some(regexp);
-        self
-    }
-
-    /// Set scope regexp filter
-    pub fn with_scope_regexp(mut self, regexp: String) -> Self {
-        self.scope_regexp = Some(regexp);
-        self
-    }
-
-    /// Set version regexp filter
-    pub fn with_version_regexp(mut self, regexp: String) -> Self {
-        self.version_regexp = Some(regexp);
-        self
-    }
-}
-
-impl AuditLogOptions {
-    /// Create new audit log options
-    pub fn new(log_type: AuditLogType) -> Self {
-        Self {
-            start_time: None,
-            end_time: None,
-            lookup: None,
-            tail: None,
-            log_type,
-        }
-    }
-
-    /// Set time range for audit logs
-    pub fn with_time_range(mut self, start: SystemTime, end: SystemTime) -> Self {
-        self.start_time = Some(start);
-        self.end_time = Some(end);
-        self
-    }
-
-    /// Set lookup string for audit logs
-    pub fn with_lookup(mut self, lookup: String) -> Self {
-        self.lookup = Some(lookup);
-        self
-    }
-
-    /// Set tail number for audit logs
-    pub fn with_tail(mut self, tail: u32) -> Self {
-        self.tail = Some(tail);
-        self
-    }
-}
-
-impl AuditLogEntry {
-    /// Create a new audit log entry
-    pub fn new(response: ListAuditLogEntriesResponse, error: Option<String>) -> Self {
-        Self { response, error }
-    }
-
-    /// Check if there's an error
-    pub fn has_error(&self) -> bool {
-        self.error.is_some()
-    }
-
-    /// Get error message if present
-    pub fn error(&self) -> Option<&str> {
-        self.error.as_deref()
-    }
 }
 
 impl BasicAuth {
@@ -243,71 +43,54 @@ impl BasicAuth {
     }
 }
 
-/// Basic auth interceptor for gRPC requests
-#[derive(Clone)]
-struct BasicAuthInterceptor {
-    auth_header: Option<MetadataValue<tonic::metadata::Ascii>>,
+struct CerbosBasicAuthInterceptor {
+    request_timeout: Duration,
+    auth_header: MetadataValue<tonic::metadata::Ascii>,
 }
 
-impl BasicAuthInterceptor {
-    fn new(creds: BasicAuth) -> Result<Self> {
-        let auth_string = format!("{}:{}", creds.username, creds.password);
-        let encoded = BASE64_STANDARD.encode(auth_string);
-        let header_value = format!("Basic {}", encoded);
-        let auth_header = MetadataValue::try_from(header_value)?;
-
-        Ok(Self { auth_header: Some(auth_header) })
-    }
-}
-
-impl Interceptor for BasicAuthInterceptor {
+impl Interceptor for CerbosBasicAuthInterceptor {
     fn call(&mut self, mut request: Request<()>) -> std::result::Result<Request<()>, Status> {
         let metadata = request.metadata_mut();
+        metadata.insert("authorization", self.auth_header.clone());
 
-        if let Some(ref auth_header) = self.auth_header {
-            metadata.insert("authorization", auth_header.clone());
-        }
-
+        request.set_timeout(self.request_timeout);
         Ok(request)
     }
 }
-
+pub struct CerbosAdminClient {
+    client: CerbosAdminServiceClient<InterceptedService<Channel, CerbosBasicAuthInterceptor>>,
+}
 impl CerbosAdminClient {
-    pub async fn new_with_credentials<S: Into<String> + Send>(
-        username: String,
-        password: String,
-        options: CerbosClientOptions<S>,
-    ) -> Result<Self> {
-        let credentials = BasicAuth::new(username, password);
-
-        let mut endpoint_builder =
-            .with_context(|| format!("Failed to create endpoint for {}", endpoint))?;
-
-        if let Some(tls_config) = options.tls_config {
-            endpoint_builder = endpoint_builder
-                .tls_config(tls_config)
-                .with_context(|| "Failed to apply TLS configuration")?;
-        }
-
-        endpoint_builder = endpoint_builder.timeout(options.timeout);
-
-        endpoint_builder = endpoint_builder
-            .user_agent(options.user_agent)
-            .with_context(|| "Failed to set user agent")?;
-
-        let channel = endpoint_builder
-            .connect()
-            .await
-            .with_context(|| format!("Failed to connect to {}", endpoint))?;
-
-        let interceptor = BasicAuthInterceptor::new(credentials)?;
-
-        // Create client with interceptor
-        let client = CerbosAdminServiceClient::with_interceptor(channel, interceptor);
+    /// Create a new Cerbos client using client options
+    pub async fn new<S>(conf: CerbosClientOptions<S>) -> Result<Self>
+    where
+        S: Into<String> + Send,
+    {
+        let basic_auth = conf
+            .admin_creds
+            .as_ref()
+            .ok_or_else(|| anyhow!("admin credentials required"))?;
+        let auth_header = Self::make_auth_header(basic_auth)?;
+        let request_timeout = conf.timeout;
+        let channel = conf.build_channel()?;
+        let client = CerbosAdminServiceClient::with_interceptor(
+            channel,
+            CerbosBasicAuthInterceptor {
+                request_timeout,
+                auth_header,
+            },
+        );
 
         Ok(Self { client })
     }
-
+    fn make_auth_header(
+        auth_creds: &BasicAuth,
+    ) -> anyhow::Result<MetadataValue<tonic::metadata::Ascii>> {
+        let auth_string = format!("{}:{}", auth_creds.username, auth_creds.password);
+        let encoded = BASE64_STANDARD.encode(auth_string);
+        let header_value = format!("Basic {}", encoded);
+        MetadataValue::try_from(header_value).with_context(|| "fail to parse metadata value")
+    }
     /// Add or update policies
     pub async fn add_or_update_policy(&mut self, policies: &PolicySet) -> Result<()> {
         policies.validate()?;
